+++ src/mesh/model/radial4.hpp
@@ -37,7 +37,7 @@ namespace model {
         }
 
     public:
-        static constexpr size_t N_NEIGHBOURS = 4;
+        static constexpr int N_NEIGHBOURS = 4;
 
         template <typename Shape>
         static std::vector<Node<Scalar, N_NEIGHBOURS>> generate(const Shape& shape,
